moved products initialization at root of app and removed repeated code/initialization from components MainDashboard, CartUpdateForm and Products
